Login endpoints

+++ server/routers/post.router.js
@@ -21,12 +21,24 @@ postRouter.get("/api/posts/:id", async (req, res) => {
 
 // CREATE NEW
 postRouter.post("/api/posts", async (req, res) => {
-  const post = await PostModel.create(req.body);
+  if (!userIsLoggedIn(req)) {
+    res.status(500).json("You need to log in before you can create post");
+    return;
+  }
+
+  const post = await PostModel.create({
+    ...req.body,
+    date: new Date().toDateString(),
+  });
   res.status(200).json(post);
 });
 
 // FIND BY ID AND UPDATE
 postRouter.put("/api/posts/:id", async (req, res) => {
+  if (!userIsLoggedIn(req)) {
+    res.status(500).json("You need to log in before you can update");
+    return;
+  }
   const post = await PostModel.findByIdAndUpdate(req.params.id, req.body);
   if (!post) {
     res.status(404).json("No post was found");
@@ -37,6 +49,10 @@ postRouter.put("/api/posts/:id", async (req, res) => {
 
 // FIND BY ID AND DELETE
 postRouter.delete("/api/posts/:id", async (req, res) => {
+  if (!userIsLoggedIn(req)) {
+    res.status(500).json("You need to log in before you can delete");
+    return;
+  }
   const post = await PostModel.findByIdAndDelete(req.params.id);
   if (!post) {
     res.status(404).json("No post was found");
@@ -44,5 +60,12 @@ postRouter.delete("/api/posts/:id", async (req, res) => {
   }
   res.status(200).json(post);
 });
+
+function userIsLoggedIn(req) {
+  if (req.session.username) {
+    return true;
+  }
+  return false;
+}
 
 export default postRouter;
